add: Lambda Origin
Set the origin to a Lambda Function using the Lambda Function URL
feature. Locking permissions down on the S3 Bucket to the Lambda
Function. Where the Lambda function permissions are locked down to
CloudFront. Which CloudFront needs to send a special HTTP header that
the code in the Lambda Function looks for to allow it CloudFront to use
the Lambda Function URL. So that if any external sources get the
function URL, they also need to know the special header to send in order
get the content.

There are better ways to secure the function URL, but this is a start.