fix: Go Lambda Code Not Loading S3 Objects
Refactored code for downloading objects from S3 and adding more logging.
Turns out the problem was with the permissions, the S3 inline policy was
not enough. Adding permissions to the Lambda IAM role fixed the issue.

Also the environment variable for setting the S3 bucket name was also
not set. Plus a few other logic bugs in the code have been fixed and
pages are loading now. However there were manual changes made in the AWS
console that need to be put into Terraform. The code is currently
loading pages from S3.