fix: CloudFront Missing A CNAME
So the domain name needs to also be added to CloudFront as an CNAME (Alt
Domain Name). That does not happen just becase you add it as an origin.

This commit will also make sure there is an alias record for each extra
alternate domain name that also points to the CloudFront distribution.
That those extra alternate domain names also are added to the CloudFront
distribution as CNAMEs.