add: CloudFront Function Viewer-Host Header
Added an function to be used with the cache behavior to copy the Host
header when the "viewer request" comes in before CloudFront redirects to
the origin. Allows a copy of the Host into another header that will not
be overwritten by CloudFront.

This seems to be a missing feature of CloudFront since it does not let
us see the original Host name the client requested in their browser.