chg: Origin and Ordered Cache Bahavior Handling
Added Terraform variable cf_additional_ordered_cache_behaviors to
handle additional ordered cache behaviors for CloudFront distributions.
Now cf_additional_origins only handles additional origins. This is
a breaking change that allows users to separately define ordered
cache behaviors and origins, enhancing the flexibility of CloudFront
configurations.
However, to tie those origins to the behaviors, the user must manually
set the origin_id in the cf_additional_ordered_cache_behaviors; which means
you need to know the names of the origins you are adding. The
key used for the addtional origins is the same as the ID used for
those origins. Its the origins that are automatically generated. You
will have to review the docs to calcualte the genereaed name to use
them in the ordered cache behaviors. This is a breaking change, as it
removed path_pattern from the cf_additional_origins aloing with a
few others that have been moved to the ordered cache behaviors
were they belong.

BREAKING CHANGE: Refactored Terraform Variables