chg: Replaced Allowed HTTP Method Variable
Replaced the allowed HTTP method variable with a boolean. You now
only have the options of Basic HTTP verbs or all HTTP verbs.
Basic is GET, HEAD, and OPTIONS, and all HTTP verbs is everything else.
This is because CloudFront only allows for 3 options:
1. GET, HEAD
2. GET, HEAD, OPTIONS
3. GET, HEAD, OPTIONS, PUT, PATCH, POST, DELETE
With this change you only get 2, or 3. As a bonus, not setting the verbs manually
for the included lambda or any additional behaviors will use which ever this
results in.