fix: Account and Region Variables
Removed the variables `aws_account` and `aws_region` from the Terraform
configuration. These were used to pass to modules and set values in
templates. The region and account set in the AWS provider will now
be used for that purpose.

These only served as confusion as they would cause your policies to
point to resources in a different account or region; when they were
set to values different than what was in the default provider.