chg: Wait For Certificate Manager Validation
Necessary to allow test to work. Currently test consistently timeout
as the validation never completes. Terraform has been ruled out as the
cause as the same issue occurs when manually creating the certificate
in the console. You can even verify the DNS records are correct and
still the validation never completes.